Added jquery-turbolinks to required gems
Chosen-Rails is a great gem. But I was running into an issue when I tried to use it with Rails 4. If you don't reload the page, Chosen would go back to being a normal select tag. So I had to add jquery-turbolinks gem to get it back to run in its normal behavior.